検索履歴リストウィジェットを追加し、カスタムドロワーで使用するように変更した
